yolov5-sunspot-hunter/yolov5-sunspot-hunter/requirements.txt
gradio==4.29.0  # 
torch
numpy
Pillow
matplotlib
opencv-python


